rubocop changed code

+++ lesson_4/lib/telemetry.rb
@@ -23,24 +23,24 @@ class Telemetry
     file_log = File.new(FILE_LOG, 'r:UTF-8')
     file_log.readlines
   end
+
   def getjson
     param_arr = []
     dev = ''
     getinfo.each do |string|
-      string = string.gsub /^$\n/, ''
+      string = string.gsub(/^$\n/, '')
       string.each_line do |str|
-      if !str.start_with?("\t") || !str.start_with?('')
-        if param_arr.empty? != true
-          @main_arr.push(dev => param_arr)
-          param_arr = []
-        end
-        dev = str[0..7].rstrip!
-        param_arr.push(
-          str.gsub(str[0..7], '').lines(':')[0].chop.gsub(' ', '_').downcase =>
-          str.gsub(str[0..7], '').lines(':')[1].chop.lstrip!.downcase
-        )
-      else
-        if str.gsub(str[0], '').lines(':')[1].nil? == false
+        if !str.start_with?("\t") || !str.start_with?('')
+          if param_arr.empty? != true
+            @main_arr.push(dev => param_arr)
+            param_arr = []
+          end
+          dev = str[0..7].rstrip!
+          param_arr.push(
+            str.gsub(str[0..7], '').lines(':')[0].chop.gsub(' ', '_').downcase =>
+            str.gsub(str[0..7], '').lines(':')[1].chop.lstrip!.downcase
+          )
+        elsif str.gsub(str[0], '').lines(':')[1].nil? == false
           param_arr.push(
             str.gsub(str[0], '').lines(':')[0].chop.gsub(' ', '_').downcase =>
             str.gsub(str[0], '').lines(':')[1].chop.lstrip!.downcase
@@ -50,19 +50,15 @@ class Telemetry
         end
       end
     end
-    end
     @main_arr
   end
-def chkfiles
-  if Dir.exist?(PATH) == false
-    Dir.mkdir(PATH)
-  else
-    if File.exist?(FILE_LOG)
-      File.delete(FILE_LOG)
-    end
-    if File.exist?(FILE_JSON)
-      File.delete(FILE_JSON)
+
+  def chkfiles
+    if Dir.exist?(PATH) == false
+      Dir.mkdir(PATH)
+    else
+      File.delete(FILE_LOG) if File.exist?(FILE_LOG)
+      File.delete(FILE_JSON) if File.exist?(FILE_JSON)
     end
   end
-end
 end
